+++ mwcp/__init__.py
@@ -4,4 +4,4 @@ from mwcp.parser import Parser
 from mwcp.parsers import register_parser_directory, iter_parsers, get_parser_descriptions
 from mwcp.reporter import Reporter
 from mwcp.resources import techanarchy_bridge
-from mwcp.resources.dispatcher import Dispatcher, ComponentParser, FileObject, UnableToParse
+from mwcp.resources.dispatcher import Dispatcher, ComponentParser, FileObject, UnableToParse, UnidentifiedFile